+++ eido/__init__.py
@@ -4,8 +4,9 @@ Project configuration
 
 from ._version import __version__
 from .const import *
+from .exceptions import *
 from .eido import *
 
-__all__ = ["validate_project", "validate_sample", "validate_config"]
 
-logmuse.init_logger(PKG_NAME)
+__all__ = ["validate_project", "validate_sample", "validate_config",
+           "read_schema", "inspect_project"]