+++ eido/const.py
@@ -1,5 +1,14 @@
 """
 Constant variables for eido package
 """
-
+LOGGING_LEVEL = "INFO"
 PKG_NAME = "eido"
+INSPECT_CMD = "inspect"
+VALIDATE_CMD = "validate"
+SUBPARSER_MSGS = {
+    VALIDATE_CMD: "Validate the PEP or its components.",
+    INSPECT_CMD: "Inspect a PEP."
+}
+PROP_KEY = "properties"
+REQUIRED_FILES_KEY = "required_files"
+FILES_KEY = "files"